+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this provider will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### :spider: Fixed
+* Fixed PK error with elasticsearch domains [#384](https://github.com/cloudquery/cq-provider-aws/pull/384).
+
 ## [v0.8.4] - 2021-12-23
 ###### SDK Version: 0.5.7