+++ CHANGELOG.md
@@ -5,9 +5,24 @@ All notable changes to this provider will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+<!-- 
 ## Unreleased
+
+### 🚀 Added
+
+### :gear: Changed
+
+### :spider: Fixed
+
+### 💥 Breaking Changes
+
+-->
+
+## [v0.8.5] - 2022-01-03
 ### :spider: Fixed
 * Fixed PK error with elasticsearch domains [#384](https://github.com/cloudquery/cq-provider-aws/pull/384).
+### :gear: Changed
+* Updated to SDK version v0.6.1
 
 ## [v0.8.4] - 2021-12-23
 ###### SDK Version: 0.5.7